Remove perf tests for 5.2
Motivation:

The performance tests have recently become quite unstable for 5.2
and require a few runs to pass. While we could broaden the slack allowed
or investigate further, it doesn't seem worth it: NIO will drop support
for 5.2 when 5.6 is released at which point we'll have to do the same.

Modifications:

- Remove the allocation counting tests for Swift 5.2

Result:

- 5.2 performance is no longer monitored
- CI is less flaky.