Provide an error when cancelling async writer
Motivation:

The `AsyncWriter` fails any pending writes with a `CancellationError` if
the writer is cancelled.

This can be misleading if the writer is cancelled because a connection
could not be established, for example.

Modifications:

- Pass an error through `AsyncWriter.cancel`

Result:

Pending writes which are failed because the writer has been cancelled
have more relevant errors.