Add async version of BenchmarkService
Motivation:

We want to provide an async/await version of the QPS Benchmark service, and be able to choose whether we want to use that or the ELF version when running the tests.

Modifications:

Added support for an async/await implementation of the Benchmark Service. This is currently not in use, but a follow-up CR will include changes to the Worker Service and other classes, and will enable choosing between the ELF and async/await implementations when running.

Result:

An async/await version of the Benchmark Service will be ready for use.