Add index docc page and fix comment links
# Motivation
We want to be able to generate proper docc pages.

# Modification
This PR adds a new index page for the `GRPC` module which is adapting the current `README`. Furthermore, it fixes a bunch of links in the doc comments so that they resolve properly.

# Result
We can now build better docc pages.

It would be great if we can convert some of the other documentations and examples over into the docc format