Auto extract IDs from headers on the server
Motivation:

It's common for service authors to extract a tracing ID from a request
header and attach that value to the logger for an RPC. This only
requires a few lines of code but must be done in each RPC. gRPC should
provide configuration to do this automatically.

Modifications:

- Add server configuration which extracts a value from headers and
  attaches it to a logger at the start of each RPC.
- Wire up the config. Add tests.

Result:

Trace IDs can be automatically extracted from headers on the server and
attached to loggers.