Avoid copies of large payloads
Motivation:

Messages are prefixed with a 5-byte header. Currently messages of all
sizes are written into a new buffer with their header. For smaller
payloads this is good: we avoid the extra allocations associated with
creating HTTP/2 frames. For large payloads the cost of the copy
outweighs the cost of extra allocations.

Modifications:

- For messages larger than 8KB emit an extra HTTP/2 DATA frame
  containing just the message header.

Result:

Better performance for large payloads.