Ignore 'NoSuchStream' errors
Motivation:

If NIOHTTP2 emits a WINDOW_UPDATE frame from a stream and the connection
is not writable then the WINDOW_UPDATE frame will not be processed
immediately. If the stream is closed before the frame is processed then
a NoSuchStream error is emitted. gRPC will close the connection if the
error is not handled, bringing down any healthy streams in the process.

Modifications:

- Ignore NoSuchStream errors

Result:

Connections aren't closed on NoSuchStream errors.