Do not check source of logs in tests
Motivation:

swift-log captures the 'source' of a log, which is usually the defining
module. In older Swift versions this was done by parsing `#file` and was
not always correct as it would just take the directory containing the
file rather than the source module: we added a check after each test in
gRPC to make sure the source was always "GRPC".

This is no longer necessary as swift-log now does the right thing.

Modifications:

- Remove the check

Result:

Tests are slightly faster.