Nil out the `onStart` callback after using it
Motivation:

The `ClientTransport` has an `onStart` callback which is called once the
stream has been established. For async calls this is used to notify the
writability manager that it may now start writing. The callback
references the call which holds the transport which holds the callback
forming a strong retain cycle and a slow memory leak. The transport
should break this cycle.

Modifications:

- nil out the `onStart` callback after it has been called.

Result:

Fixes a leak