Tool that generates serialised file descriptor protos
Motivation:

In order to implement the reflection service we need to be able to generate serialised file descriptor protos of the
services and messages that a server offers.

Modifications:

Added an option for protoc-gen-grpc-swift that enables generating a binary file containing a base64 encoded representation of the  serialized file descriptor proto of the specified proto file.

Result:

This change enables the generation of serialised file descriptor protos that will be useful in the implementation of the reflection service.