Add transport protocols and surrounding types
Motivation:

The transport protocols provide a lower-level abstraction for different
conection protocols. For example, there will be an HTTP/2 transport and
in the future, when an implementation arises, a separate HTTP/3 transport.
This change adds these interfaces and a handful of related types which
surround the transport protocols.

Modifications:

- Add RPC parts
- Add RPC stream composed of streams inbound and outbound RPC parts
- Add async sequence wrapper
- Add writer protocol and wrapper
- Add various currency types
- Tests

Results:

Transport protocols and related types are in place