Generated reflection service code
Motivation:

In order to develop the Reflection Service, the specific proto file should be included in the module and
the service code generated from it.

Modifications:

Created the module for the Reflection Service, added the proto file, updated the Package.swift
to include the module, updated the Makefile to inlude targets for generating the server code,
created the ReflectionProvider.

Result:

The Reflection Service can now be implemented.