Add a broadcast async sequence
Motivation:

To support retries and hedging we need a way to buffer elements over
time that can support multiple consumers concurrently and allows for
consumers to start consuming after some elements have been produced.

An `AsyncSequence` fits this quite naturally but we don't yet have a
general purpose implementat that fits this requirement. This change
adds `BroadcastAsyncSequence` which isn't a general purpose async
sequence but instead is tailored to the needs of grpc for hedging and
retries. This means it supports a low number of concurrent iterators and
maintains a limited size internal buffer and drops the slowest consumers
when the buffer becomes full.

Modifications:

- Add a `BroadcastAsyncSequence` and tests
- Made a bunch of things inlinable/usableFromInline which necessitated a
  switch from `@_spi(Testing)` to `@testable` imports.
- Rename the 'Stream' directory to 'Streaming'

Result:

- `BroadcastAsyncSequence` can be used to implement retries and hedging.