Add a handful of utilities
Motivation:

The client rpc executor makes use of a bunch of utilities. Since it will
be a reasonably large PR, in order to make it slightly less large, I'd
like to get some of the utilities reviewed separately. Since most are
too small to be worth reviewing individually this change includes a
few unrelated utilities.

Modifications:

- Add `UnsafeTransfer`
- Adds an optional `cause` error to `RPCError`
- Add extensions to `Metadata` for setting/parsing a few gRPC specific
  metadata fields
- Add extensions to `Result` for working with `async` closures and
  casting errors to a known type
- Add a type-erased closable writer similar to the type-erased writer

Result:

A few handy helpers are in place and the rpc executor PR will be a
little smaller.