Basic client RPC executor
Motivation:

Clients can execute RPCs in a few different ways: straight up as a
single attempt, with retries, or with hedging. This is done by the
`ClientRPCExecutor`, the bones of which are added in this PR.

At a high level, the executor takes a transport, request, serializer,
deserializer and response handler and executes the request against a
transport and executes the response handler if a response is received or
synthesized locally. The executor only returns once the response has
been handled.

Modifications:

- Add the `ClientRPCExecutor` and the one-shot implementation (hedging
  and retries will follow later)
- The `ClientRPCExecutor` uses a `ClientStreamExecutor` which deals in
  serialized streams
- Add a testing harness, which includes type erased transports (and a
  basic in process transport for testing) and different server
  behaviours.

Result:

Can execute one-shot requests.