Created README for the Reflection Service
Motivation:

To help users set up a Server that is implementing the Reflection Service.

Modifications:

Added a step by step tutorial on how to set up the example server and run GRPCurl
in order to test the Reflection Service.

Result:

Users will get information on how to set up the Reflection Service for a Server.