Changed the name of the package of the reflection.proto
Motivation:

The name was not complete and GRPCurl's requests for the Reflection Service
resulted in error.

Modifications:

Changed the name of the package, regenerated the Service code and
changed the type of the variables accordingly, in the Service implementation
and in the tests.

Result:

GRPCurl will be able to make requests to Servers implementing the
Reflection Service.