Add the server RPC executor.
Motivation:

The server needs to handle accepted RPC streams by turning them into
requests and letting a user provided handler handle the request, turning
it into the response. The server executor is then responsible for
handling the response and writing it back to the client.

Modifications:

- Add the server RPC executor
- Add testing Utilities and tests for the server executor

Result:

Server RPCs can be handled