Add the router and service protocol
Motivation:

The transport layer accepts streams from the client and transforms them
into requests which are handled by services. Services, written by users,
need a way to tell the server which streams should be handled by which
methods.

Modifications:

- Add the RPC router which routes streams to "handlers"
- Add the RegistrableRPCService protocol which services can conform to
  to register routes with a router.

The `RegistrableRPCService` is typically not implemented by users and
instead will be implemented in the generated code. However, it builds on
public API offered by the router.

Result:

We have interfaces which allow streams to be routed to a service
handler.