Bring over structured Swift syntax from OpenAPI
Motivation:

We’ll need to be able to represent code to generate as structured Swift types for the new Codegen library.
This is effectively what swift-syntax provides but we can’t (at time of writing) take it as a dependency
because of its unstable API. Swift OpenAPI Generator has already provided a number of these types as its StructuedSwiftRepresentation.

Modifications:

- Copied StructuredSwiftRepresentation.swift inside the Codegen library module
- Mentioned the usage in NOTICES.txt

Result:

Codegen library will have a representation of the code to generate as structured Swift types.