Added TypeName and TypeUsage from OpenAPI Generator.
Motivation:

The types are used by the code renderer that will be brought over from OpenAPI Generator
to the new gRPC Swift CodeGenLib.

Modifications:

Copied the TypeName, TypeUsage and some TypeName extensions that will be used
in the context of gRPC, and modified TypeName so it represents only the Swift path
of a type.

Result:

We will be able to bring over the Code Renderer and we will have types representations that can be used
by the service translator.