[CodeGenLib] Translator for server code
Motivation:

In the generated server code for a service described in a Source IDL file
we want to have:
- a protocol defining all the methods as bidirectional streaming, conforming to `RegistrableRPCService`
- an extension for the streaming protocol that implements the `registerRPCs`
function that is required by the `RegistrableRPCService` conformance
- a service protocol that defines the methods of the service, as they are
descibed in the Source IDL (unary/input-streaming/output-streaming/bidirectional-streaming) that conforms to the streaming protocol.
- the extension of the service protocol that implements the bidirectional
streaming methods required by the conformance to the streaming protocol, using
calls to the service methods (in case a service method is bidirectional
streaming), it doesn't need to be reimplemented.

Modifications:

- Created the ServerCodeTranslator that conforms to SpecializedTranslator and
contains all the methods for creating the representation of the generated code
that we want using StructuredSwiftRepresentation format.
- Implemented snippet tests that test the ServerCodeTranslator.
- added the inout parameter type as a case of ExistingTypeDescription in StructuredSwiftRepresentation, as it is needed by the `registerRPCs` function.

Result:

Server code can now be generated for a CodeGenerationRequest object.